Rename trace/constants.py to trace/core.py and consolidate shared structures.
- Move TryExceptBlock and ChainLink dataclasses from chain_analysis into core.
- Consolidate the duplicate Range namedtuple in syntaxerror.py into core.
- Add frozenset constants for quote/triple/prefix/emphasis/relevance checks.
- Replace ad-hoc tuples/lists with the new frozensets in fragments, digest,
  syntaxerror, order, finalize, html, and tty.
- Update tests that imported from tracerite.trace.constants to core.
